requirements.txt
blinker==1.6.2
click==8.1.6
DateTime==5.2
dnspython==2.4.1
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3
Werkzeug==2.3.6
zipp==3.16.2
zope.interface==6.0
